refactor: typescript file naming
Usual TypeScript file naming for EventListenerStore and DropdownMenuStatic.
