+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-dropdown-menu-static.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import {DropdownMenuAlignment} from "../tobago-dropdown-menu";
+import {DropdownMenuAlignment} from "./tobago-dropdown-menu";
 
 export class DropdownMenuStatic {
