+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-file.ts
@@ -18,7 +18,7 @@
 import {Overlay} from "./tobago-overlay";
 import {OverlayType} from "./tobago-overlay-type";
 import {Css} from "./tobago-css";
-import {EventListenerStore} from "./util/EventListenerStore";
+import {EventListenerStore} from "./tobago-event-listener-store";
 
 export class File extends HTMLElement {
   private listeners: EventListenerStore = new EventListenerStore();
