+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-in.ts
@@ -17,7 +17,7 @@
 
 import {Focus} from "./tobago-focus";
 import {Suggest} from "./tobago-suggest";
-import {EventListenerStore} from "./util/EventListenerStore";
+import {EventListenerStore} from "./tobago-event-listener-store";
 
 export class In extends HTMLElement {
   private listeners: EventListenerStore = new EventListenerStore();
