+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-paginator.ts
@@ -19,7 +19,7 @@ import {Key} from "./tobago-key";
 import {Overlay} from "./tobago-overlay";
 import {OverlayType} from "./tobago-overlay-type";
 import {Page} from "./tobago-page";
-import {EventListenerStore} from "./util/EventListenerStore";
+import {EventListenerStore} from "./tobago-event-listener-store";
 
 export class TobagoPaginator extends HTMLElement {
   private listeners: EventListenerStore = new EventListenerStore();
