+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-select-one-radio.ts
@@ -17,7 +17,7 @@
 
 import {Focus} from "./tobago-focus";
 import {Key} from "./tobago-key";
-import {EventListenerStore} from "./util/EventListenerStore";
+import {EventListenerStore} from "./tobago-event-listener-store";
 
 class SelectOneRadio extends HTMLElement {
   private listeners: EventListenerStore = new EventListenerStore();
