+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-split-layout.ts
@@ -17,7 +17,7 @@
 
 import {Page} from "./tobago-page";
 import {Css} from "./tobago-css";
-import {EventListenerStore} from "./util/EventListenerStore";
+import {EventListenerStore} from "./tobago-event-listener-store";
 
 class SplitLayout extends HTMLElement {
   private listeners: EventListenerStore = new EventListenerStore();
