+++ tobago-theme/tobago-theme-standard/src/main/ts/tobago-tree-node.ts
@@ -18,7 +18,7 @@
 import {Sheet} from "./tobago-sheet";
 import {Tree} from "./tobago-tree";
 import {Css} from "./tobago-css";
-import {EventListenerStore} from "./util/EventListenerStore";
+import {EventListenerStore} from "./tobago-event-listener-store";
 
 export class TreeNode extends HTMLElement {
   private listeners: EventListenerStore = new EventListenerStore();
